Fix Bug in Counter, found in https://github.com/abergs/fido2-net-lib. We must check the Endian :-/.

+++ src/Sternzeit.Server/Services/LoginExpectations.cs
@@ -12,7 +12,7 @@ namespace Sternzeit.Server.Services
         public string[] Origins { get; set; }
         public byte[] RelayingPartyHash { get; set; }
 
-        public int Counter { get; set; }
+        public long Counter { get; set; }
 
         public PublicKey PublicUserKey { get; set; }
     }
